src/requirements.txt
asn1crypto==1.5.1
async-exit-stack==1.0.1
async-generator==1.10
attrs==23.1.0
base58==1.0.3
cffi==1.16.0
cryptography==41.0.4
ecdsa==0.18.0
fastecdsa==1.7.5
idna==3.4
leveldb==0.201
lru-dict==1.2.0
multiaddr==0.0.9
mypy==1.6.0
mypy-extensions==1.0.0
netaddr==0.9.0
outcome==1.2.0
protobuf==3.20.3
pycparser==2.21
pycryptodome==3.19.0
pymultihash==0.8.2
PyNaCl==1.3.0
pyzmq==25.1.1
redis==5.0.1
rpcudp==3.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.66.1
trio==0.22.2
typing_extensions==4.8.0
u-msgpack-python==2.8.0
varint==1.0.2
